additional styling changes... started making finished/done button functionality
