+++ src/react-modules/FileUpload.js
@@ -67,14 +67,11 @@ function FileUploadFull({
             </>
           )}
 
-          <p className=" text-center mb-2 text-xs text-zinc-500 dark:text-zinc-400 lowercase">
+          <p className=" text-center text-xs text-zinc-500 dark:text-zinc-400 lowercase">
             <span className="font-bold text-zinc-300 dark:text-zinc-200">
               Click to upload
             </span>{" "}
             or drag and drop
-          </p>
-          <p className="text-xs text-zinc-500 dark:text-zinc-400 dark:text-zinc-400">
-            (SVG, PNG, JPEG, JPG)
           </p>
         </div>
