+++ src/react-modules/character-screen/CharacterScreen.js
@@ -1,11 +1,9 @@
 import { NewCharacterButton } from "./NewCharacterButton";
-import { TextInput, TextAreaInput } from "./../TextInput";
-import { FileUploadFull, FileUploadBar } from "./../FileUpload";
 import "./character-styles/character-screen.css";
-import { useState } from "react";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faUser } from "@fortawesome/free-solid-svg-icons";
-import {CharacterInfo} from "./CharacterInfo"
+import { useState, useEffect } from "react";
+import { CharacterInfo } from "./CharacterInfo";
+import { CharacterEdit } from "./CharacterEdit";
+import { CharacterList } from "./CharacterList";
 
 function CharacterScreen({
   showcaseJSON,
@@ -15,126 +13,94 @@ function CharacterScreen({
   handleJSONUpdate,
 }) {
 
-
-  const handleClick = (e) => {
-    setSelectedCharacter(e.currentTarget.value);
-  };
-
-  const handleRemove = (e, i) => {
-    if (showcaseJSON.personas.length == 1) return;
-
-    // Prevent out of bounds selected character
-    if (
-      (selectedCharacter == i ||
-        showcaseJSON.personas.length - 1 == selectedCharacter) &&
-      selectedCharacter != 0
-    ) {
-      setSelectedCharacter(selectedCharacter - 1);
-    }
-
-    setShowcaseJSON((json) => {
-      json["personas"].splice(i, 1);
-    });
-  };
-
   const [editMode, setEditMode] = useState(false);
 
+  // Seperately update a mini version of the json, containing only the fields for this page
+  const [localJSON, setLocalJSON] = useState(); 
+
+  // Change this mini version of the json, when the character changes
+  useEffect(() => {
+    setLocalJSON({
+        "name":showcaseJSON.personas[selectedCharacter].name,
+        "type":showcaseJSON.personas[selectedCharacter].type,
+        "description":showcaseJSON.personas[selectedCharacter].description,
+      })
+  }, [selectedCharacter]);
+
+  // To-do: Create a function similar to handleJSONUpdate in App.js
+  // Then, instead of passing the actual JSON to the text fields in <CharacterEdit>, pass the mini json and the mini handler
+
+  // To-do: Impliment a save handler. When clicking save, send the mini JSON to the real, full JSON file
+
   return (
-    <div className="justify-center items-center flex content-center">
-      <div className="flex p-3 w-2/5 justify-center items-center flex-col">
-        <div className="flex w-full">
-          <div>
-          <h2 className="text-2xl text-white text-start w-full font-bold">
-            Select Your Character
-          </h2>
-          <p className="w-full">Select a character below or create a new one.</p>
-          </div>
-          {/* ADD BUTTON */}
-          <div className="ml-auto m-5">
-            <NewCharacterButton
-              showcaseJSON={showcaseJSON}
-              setShowcaseJSON={setShowcaseJSON}
-              setSelectedCharacter={setSelectedCharacter}
-            />
-          </div>
-        </div>
-
-        <div className="grid grid-cols-3">
-          {showcaseJSON.personas.map((person, index) => (
-            <button value={index} key={index} onClick={handleClick}>
-              <div>
-                <div
-              
-                  
-                  className={`character-circle flex items-center justify-center p-3 m-3  ${
-                    selectedCharacter == index ? "selected-item" : ""
-                  }`}
-                >
-                  <FontAwesomeIcon icon={faUser} />
-                </div>
-                <div className="p-2">
-                  <p className="text-center">{person.name}</p>
-                </div>
-              </div>
-            </button>
-          ))}
-        </div>
-      </div>
-
-      <div className="highlight-container w-2/5 rounded p-3">
-        {editMode ? (
-          // Toggling edit mode
-          <>
-            <h2 className="text-1xl text-neutral-700 dark:text-white font-bold mb-5">
-              EDIT CHARACTER
-            </h2>
-            <div className="grid grid-cols-2 w-full">
-              <TextInput
-                label={"Name:"}
-                personaIndex={selectedCharacter}
-                element={["name"]}
-                handleJSONUpdate={handleJSONUpdate}
+    <>
+      <div className="justify-center items-center flex content-center">
+        <div className="flex p-3 w-2/5 justify-center items-center flex-col">
+          <div className="flex w-full">
+            <div>
+              <h2 className="text-2xl text-white text-start w-full font-bold">
+                Select Your Character
+              </h2>
+              <p className="w-full">
+                Select a character below or create a new one.
+              </p>
+            </div>
+            {/* ADD BUTTON */}
+            <div className="ml-auto m-5">
+              <NewCharacterButton
                 showcaseJSON={showcaseJSON}
-              />
-              <TextInput
-                label={"Role:"}
-                personaIndex={selectedCharacter}
-                element={["type"]}
-                handleJSONUpdate={handleJSONUpdate}
-                showcaseJSON={showcaseJSON}
+                setShowcaseJSON={setShowcaseJSON}
+                setSelectedCharacter={setSelectedCharacter}
               />
             </div>
-            <TextAreaInput
-              label={"Page Description:"}
-              personaIndex={selectedCharacter}
-              element={["description"]}
+          </div>
+
+          <CharacterList
+            setEditMode={setEditMode}
+            showcaseJSON={showcaseJSON}
+            selectedCharacter={selectedCharacter}
+            setSelectedCharacter={setSelectedCharacter}
+          />
+        </div>
+
+        <div className="highlight-container w-2/5 rounded p-3">
+          {editMode ? (
+            // Toggling edit mode
+            <CharacterEdit
+              selectedCharacter={selectedCharacter}
               handleJSONUpdate={handleJSONUpdate}
               showcaseJSON={showcaseJSON}
             />
-            <div className="grid grid-cols-3 w-full">
-              <FileUploadFull
-                text={"Body Image"}
-                personaIndex={selectedCharacter}
-                element={["image"]}
-                handleJSONUpdate={handleJSONUpdate}
-                showcaseJSON={showcaseJSON}
-              />
+          ) : (
+            <CharacterInfo
+              setShowcaseJSON={setShowcaseJSON}
+              showcaseJSON={showcaseJSON}
+              setSelectedCharacter={setSelectedCharacter}
+              selectedCharacter={selectedCharacter}
+              setEditMode={setEditMode}
+            />
+          )}
+          {/* Save or cancel button */
+          editMode ? (
+          <div className="flex flex-cols mx-5 my-3 justify-end space-x-4">
+            <button
+              className="p-1 w-20 hover:underline uppercase"
+              onClick={() => setEditMode(false)}
+            >
+              Cancel
+            </button>
 
-              <FileUploadFull
-                text={"Avatar Image"}
-                personaIndex={selectedCharacter}
-                element={["revocationInfo", 0]}
-                handleJSONUpdate={handleJSONUpdate}
-                showcaseJSON={showcaseJSON}
-              />
-            </div>
-          </>
-        ) : (
-          <CharacterInfo showcaseJSON={showcaseJSON} selectedCharacter={selectedCharacter} handleRemove={handleRemove} setEditMode={setEditMode}/>
-        )
-        }
+            <button
+              className="p-1 w-20 button-dark hover:bg-neutral-600"
+              onClick={() => setEditMode(false)}
+            >
+              SAVE
+            </button>
+          </div>) : null
+          }
+        </div>
       </div>
-    </div>
+    </>
   );
 }
